Drop support for passing an organization's name
This is a bit tricky to do if resolving the org's name is done at
request time, because it makes for a much more expensive
token-creating request.

There may be a good way to implement this by adding some caching,
though care should be taken to verify the name == installation ID
to be robust against renames.

I'm not deleting this feature because I don't like it, but because
retaining the implementation was tricky, and retaining the code
without actually using it it made code coverage go below 90%.